pyproject.toml
[project]
name = "wrist"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.4",
    "livekit-plugins-anthropic>=1.5",
    "livekit-plugins-noise-cancellation>=0.2",
    "python-dotenv",
    "anthropic>=0.42",
    "livekit-plugins-deepgram>=1.5.0",
    "exa-py>=1.0",
    "httpx>=0.27",
]
